Moves mapCategories higher up in the folder structure; adds check for category mismatch; adds docblocks
